vmm_tests: Reorganize multiarch.rs
